refactored objc template methods
