+++ signals/generators/ios/swift_template.py
@@ -2,7 +2,7 @@ import shutil
 from signals.generators.base.base_template import BaseTemplate
 from signals.generators.ios.conversion import sanitize_field_name, get_proper_name
 from signals.generators.ios.parameters import SwiftParameter
-from signals.generators.ios.template_methods import iOSTemplateMethods
+from signals.generators.ios.ios_template_methods import iOSTemplateMethods
 from signals.parser.api import GetAPI
 from signals.parser.fields import Field
 
@@ -32,36 +32,6 @@ class SwiftTemplate(BaseTemplate):
 
 class SwiftTemplateMethods(iOSTemplateMethods):
     @staticmethod
-    def attribute_mappings(fields):
-        attribute_mapping_string = ""
-        for index, field in enumerate(fields):
-            leading_comma = '' if index == 0 else ', '
-            swift_variable_name = get_proper_name(field.name)
-            attribute_mapping_string += '{}"{}": "{}"'.format(leading_comma, field.name, swift_variable_name)
-        return attribute_mapping_string
-
-    @staticmethod
-    def create_parameter_signature(parameters):
-        method_parts = []
-        for index, method_field in enumerate(parameters):
-            swift_variable_name = get_proper_name(method_field.name)
-            parameter_signature = "{}: {}".format(swift_variable_name, method_field.swift_type)
-
-            method_parts.append(parameter_signature)
-
-        return ", ".join(method_parts)
-
-    @staticmethod
-    def key_path(api):
-        key_path_string = 'nil'
-        if hasattr(api, 'resource_type'):
-            key_path_string = 'nil' if api.resource_type == GetAPI.RESOURCE_DETAIL else '"results"'
-        elif isinstance(api, GetAPI) and ':id' not in api.url_path:
-            # Get requests with an ID only return 1 object, not a list of results
-            key_path_string = '"results"'
-        return key_path_string
-
-    @staticmethod
     def method_parameters(api):
         parameters = []
 
@@ -86,3 +56,33 @@ class SwiftTemplateMethods(iOSTemplateMethods):
         ])
 
         return SwiftTemplateMethods.create_parameter_signature(parameters)
+
+    @staticmethod
+    def key_path(api):
+        key_path_string = 'nil'
+        if hasattr(api, 'resource_type'):
+            key_path_string = 'nil' if api.resource_type == GetAPI.RESOURCE_DETAIL else '"results"'
+        elif isinstance(api, GetAPI) and ':id' not in api.url_path:
+            # Get requests with an ID only return 1 object, not a list of results
+            key_path_string = '"results"'
+        return key_path_string
+
+    @staticmethod
+    def attribute_mappings(fields):
+        attribute_mapping_string = ""
+        for index, field in enumerate(fields):
+            leading_comma = '' if index == 0 else ', '
+            swift_variable_name = get_proper_name(field.name)
+            attribute_mapping_string += '{}"{}": "{}"'.format(leading_comma, field.name, swift_variable_name)
+        return attribute_mapping_string
+
+    @staticmethod
+    def create_parameter_signature(parameters):
+        method_parts = []
+        for index, method_field in enumerate(parameters):
+            swift_variable_name = get_proper_name(method_field.name)
+            parameter_signature = "{}: {}".format(swift_variable_name, method_field.swift_type)
+
+            method_parts.append(parameter_signature)
+
+        return ", ".join(method_parts)
